Add flag to turn of csv export

+++ multiscale_benchmark/2022_09_hackathon/Biodynamo/unit_test_diffusion/run.sh
@@ -31,5 +31,5 @@ fi
 bdm run
 
 # Move the results to the correct folder
-mv output/concentration.csv $DIR/results
 mv output/diffusion-small/metadata $DIR/results
+[ -f output/concentration.csv ] && mv output/concentration.csv $DIR/results
